+++ apps/decodex/src/orchestrator/tests/operator/status/running_lanes/ghost_lane/cleanup_blockers.rs
@@ -1,442 +1,3 @@
-use crate::orchestrator::tests::operator::status::running_lanes::{
-	self, Connection, FakeTracker, ProtocolActivityMarker, StateStore, fs, orchestrator, process,
-	state,
-};
-
-#[test]
-fn live_operator_status_allows_ghost_recovery_when_worktree_mapping_path_is_missing() {
-	let (_temp_dir, config, workflow) = running_lanes::temp_project_layout();
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let tracker = FakeTracker::new(Vec::new());
-	let missing_worktree_path = config.worktree_root().join("PUB-012");
-
-	state_store
-		.record_run_attempt("run-12", "PUB-012", 1, "running")
-		.expect("run attempt should record");
-	state_store
-		.upsert_lease("pubfi", "PUB-012", "run-12", "In Progress")
-		.expect("lease should record");
-	state_store
-		.upsert_worktree(
-			"pubfi",
-			"PUB-012",
-			"x/pubfi-pub-012",
-			&missing_worktree_path.display().to_string(),
-		)
-		.expect("stale worktree mapping should record");
-
-	let snapshot = orchestrator::build_live_operator_status_snapshot(
-		&tracker,
-		&config,
-		&workflow,
-		&state_store,
-		10,
-	)
-	.expect("snapshot should build");
-	let run = snapshot.current_lanes.first().expect("ghost current lane should be visible");
-
-	assert_eq!(run.ownership_state, "ghost_lane");
-	assert_eq!(run.policy_state, "runtime_recovery_required");
-	assert_eq!(run.lane_control_next_action, "run_ghost_lane_recovery");
-	assert!(run.lane_control_conditions.contains(&String::from("worktree_mapping_path_missing")));
-	assert!(run.lane_control_conditions.contains(&String::from("worktree_missing")));
-	assert!(!run.lane_control_conditions.contains(&String::from("retained_worktree_present")));
-}
-
-#[test]
-fn live_operator_status_blocks_missing_issue_ghost_cleanup_when_retained_worktree_exists() {
-	let (_temp_dir, config, workflow) = running_lanes::temp_project_layout();
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let tracker = FakeTracker::new(Vec::new());
-
-	fs::create_dir_all(config.worktree_root().join("PUB-012"))
-		.expect("retained worktree directory should exist");
-
-	state_store
-		.record_run_attempt("run-12", "PUB-012", 1, "running")
-		.expect("run attempt should record");
-	state_store
-		.upsert_lease("pubfi", "PUB-012", "run-12", "In Progress")
-		.expect("lease should record");
-
-	let snapshot = orchestrator::build_live_operator_status_snapshot(
-		&tracker,
-		&config,
-		&workflow,
-		&state_store,
-		10,
-	)
-	.expect("snapshot should build");
-	let run = snapshot.current_lanes.first().expect("current lane should be visible");
-
-	assert_eq!(run.ownership_state, "retained_attention");
-	assert_eq!(run.policy_state, "runtime_recovery_blocked");
-	assert_eq!(run.lane_control_next_action, "inspect_missing_issue_runtime_recovery_blockers");
-	assert!(run.needs_attention);
-	assert!(!run.counts_as_running);
-	assert!(run.lane_control_conditions.contains(&String::from("retained_worktree_present")));
-}
-
-#[test]
-fn live_operator_status_blocks_missing_issue_ghost_cleanup_when_control_channel_row_exists() {
-	let (temp_dir, config, workflow) = running_lanes::temp_project_layout();
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let tracker = FakeTracker::new(Vec::new());
-	let channel_path = temp_dir.path().join("missing-control-channel.json");
-
-	state_store
-		.record_run_attempt("run-12", "PUB-012", 1, "running")
-		.expect("run attempt should record");
-	state_store
-		.upsert_lease("pubfi", "PUB-012", "run-12", "In Progress")
-		.expect("lease should record");
-	state_store
-		.publish_run_control_channel_for_active_attempt("run-12", 1, &channel_path, "local_file")
-		.expect("control channel row should publish");
-
-	let snapshot = orchestrator::build_live_operator_status_snapshot(
-		&tracker,
-		&config,
-		&workflow,
-		&state_store,
-		10,
-	)
-	.expect("snapshot should build");
-	let run = snapshot.current_lanes.first().expect("current lane should be visible");
-
-	assert_eq!(run.ownership_state, "retained_attention");
-	assert_eq!(run.policy_state, "runtime_recovery_blocked");
-	assert_eq!(run.lane_control_next_action, "inspect_missing_issue_runtime_recovery_blockers");
-	assert!(run.needs_attention);
-	assert!(!run.counts_as_running);
-	assert!(run.lane_control_conditions.contains(&String::from("control_channel_file_missing")));
-	assert!(run.lane_control_conditions.contains(&String::from("control_channel_present")));
-}
-
-#[test]
-fn ghost_lane_cleanup_status_blockers_treat_invalid_local_issue_id_as_missing_issue() {
-	let (_temp_dir, config, workflow) = running_lanes::temp_project_layout();
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let tracker = FakeTracker::with_refresh_error(
-		Vec::new(),
-		"Linear GraphQL request failed: Argument Validation Error",
-	);
-
-	state_store
-		.record_run_attempt("run-12", "PUB-012", 1, "running")
-		.expect("run attempt should record");
-	state_store
-		.upsert_lease("pubfi", "PUB-012", "run-12", "In Progress")
-		.expect("lease should record");
-
-	let blockers = orchestrator::ghost_lane_cleanup_status_blockers(
-		&tracker,
-		&config,
-		&workflow,
-		&state_store,
-		"PUB-012",
-		"run-12",
-	)
-	.expect("invalid local issue id should be classified as a missing tracker issue");
-
-	assert!(blockers.is_empty(), "missing issue with no live evidence should allow cleanup");
-}
-
-#[test]
-fn ghost_lane_cleanup_status_blockers_preserve_live_blockers_after_invalid_issue_id_lookup() {
-	let (temp_dir, config, workflow) = running_lanes::temp_project_layout();
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let tracker = FakeTracker::with_refresh_error(
-		Vec::new(),
-		"Linear GraphQL request failed: Argument Validation Error",
-	);
-	let channel_path = temp_dir.path().join("missing-control-channel.json");
-
-	state_store
-		.record_run_attempt("run-12", "PUB-012", 1, "running")
-		.expect("run attempt should record");
-	state_store
-		.upsert_lease("pubfi", "PUB-012", "run-12", "In Progress")
-		.expect("lease should record");
-	state_store
-		.publish_run_control_channel_for_active_attempt("run-12", 1, &channel_path, "local_file")
-		.expect("control channel row should publish");
-
-	let blockers = orchestrator::ghost_lane_cleanup_status_blockers(
-		&tracker,
-		&config,
-		&workflow,
-		&state_store,
-		"PUB-012",
-		"run-12",
-	)
-	.expect("invalid local issue id should still run local safety checks");
-
-	assert!(blockers.contains(&String::from("control_channel_present")));
-	assert!(blockers.contains(&String::from("control_channel_file_missing")));
-}
-
-#[test]
-fn ghost_lane_cleanup_status_blockers_do_not_hide_validation_error_for_server_issue_id() {
-	let (_temp_dir, config, workflow) = running_lanes::temp_project_layout();
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let tracker = FakeTracker::with_refresh_error(
-		Vec::new(),
-		"Linear GraphQL request failed: Argument Validation Error",
-	);
-	let issue_id = "00000000-0000-0000-0000-000000000012";
-
-	state_store
-		.record_run_attempt("run-12", issue_id, 1, "running")
-		.expect("run attempt should record");
-	state_store
-		.upsert_lease("pubfi", issue_id, "run-12", "In Progress")
-		.expect("lease should record");
-
-	let error = orchestrator::ghost_lane_cleanup_status_blockers(
-		&tracker,
-		&config,
-		&workflow,
-		&state_store,
-		issue_id,
-		"run-12",
-	)
-	.expect_err("server issue id validation errors must remain tracker failures");
-
-	assert!(error.to_string().contains("Argument Validation Error"));
-}
-
-#[test]
-fn ghost_lane_cleanup_status_blockers_reject_live_process_evidence() {
-	let (_temp_dir, config, workflow) = running_lanes::temp_project_layout();
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let tracker = FakeTracker::new(Vec::new());
-	let worktree_path = config.worktree_root().join("PUB-012");
-
-	fs::create_dir_all(&worktree_path).expect("worktree path should exist");
-
-	state_store
-		.record_run_attempt("run-12", "PUB-012", 1, "running")
-		.expect("run attempt should record");
-	state_store
-		.upsert_lease("pubfi", "PUB-012", "run-12", "In Progress")
-		.expect("lease should record");
-	state_store
-		.upsert_worktree(
-			"pubfi",
-			"PUB-012",
-			"x/pubfi-pub-012",
-			&worktree_path.display().to_string(),
-		)
-		.expect("worktree should record");
-
-	state::write_run_activity_marker_for_process(&worktree_path, "run-12", 1, process::id())
-		.expect("live process marker should write");
-
-	let blockers = orchestrator::ghost_lane_cleanup_status_blockers(
-		&tracker,
-		&config,
-		&workflow,
-		&state_store,
-		"PUB-012",
-		"run-12",
-	)
-	.expect("cleanup status blockers should load");
-
-	assert!(blockers.contains(&String::from("process_alive")));
-	assert!(blockers.contains(&String::from("retained_worktree_present")));
-}
-
-#[test]
-fn ghost_lane_cleanup_status_blockers_reject_active_thread_evidence() {
-	let (_temp_dir, config, workflow) = running_lanes::temp_project_layout();
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let tracker = FakeTracker::new(Vec::new());
-	let worktree_path = config.worktree_root().join("PUB-012");
-
-	fs::create_dir_all(&worktree_path).expect("worktree path should exist");
-
-	state_store
-		.record_run_attempt("run-12", "PUB-012", 1, "running")
-		.expect("run attempt should record");
-	state_store
-		.upsert_lease("pubfi", "PUB-012", "run-12", "In Progress")
-		.expect("lease should record");
-	state_store
-		.upsert_worktree(
-			"pubfi",
-			"PUB-012",
-			"x/pubfi-pub-012",
-			&worktree_path.display().to_string(),
-		)
-		.expect("worktree should record");
-
-	state::write_run_thread_status_marker(
-		&worktree_path,
-		"run-12",
-		1,
-		Some("thread-12"),
-		Some("turn-12"),
-		"active",
-		&[String::from("waitingOnApproval")],
-	)
-	.expect("active thread marker should write");
-
-	let blockers = orchestrator::ghost_lane_cleanup_status_blockers(
-		&tracker,
-		&config,
-		&workflow,
-		&state_store,
-		"PUB-012",
-		"run-12",
-	)
-	.expect("cleanup status blockers should load");
-
-	assert!(blockers.contains(&String::from("thread_active")));
-	assert!(blockers.contains(&String::from("retained_worktree_present")));
-}
-
-#[test]
-fn ghost_lane_cleanup_status_blockers_do_not_persist_marker_thread_identity() {
-	let (temp_dir, config, workflow) = running_lanes::temp_project_layout();
-	let state_path = temp_dir.path().join("runtime.sqlite3");
-	let state_store = StateStore::open(&state_path).expect("state store should open");
-	let tracker = FakeTracker::new(Vec::new());
-	let worktree_path = config.worktree_root().join("PUB-012");
-
-	fs::create_dir_all(&worktree_path).expect("worktree path should exist");
-
-	state_store
-		.record_run_attempt("run-12", "PUB-012", 1, "running")
-		.expect("run attempt should record");
-	state_store
-		.upsert_lease("pubfi", "PUB-012", "run-12", "In Progress")
-		.expect("lease should record");
-	state_store
-		.upsert_worktree(
-			"pubfi",
-			"PUB-012",
-			"x/pubfi-pub-012",
-			&worktree_path.display().to_string(),
-		)
-		.expect("worktree should record");
-
-	state::write_run_thread_status_marker(
-		&worktree_path,
-		"run-12",
-		1,
-		Some("thread-12"),
-		Some("turn-12"),
-		"active",
-		&[String::from("waitingOnApproval")],
-	)
-	.expect("active thread marker should write");
-
-	let blockers = orchestrator::ghost_lane_cleanup_status_blockers(
-		&tracker,
-		&config,
-		&workflow,
-		&state_store,
-		"PUB-012",
-		"run-12",
-	)
-	.expect("cleanup status blockers should load");
-	let connection = Connection::open(&state_path).expect("sqlite should open");
-	let (thread_id, turn_id): (Option<String>, Option<String>) = connection
-		.query_row(
-			"SELECT thread_id, turn_id FROM run_attempts WHERE run_id = 'run-12'",
-			[],
-			|row| Ok((row.get(0)?, row.get(1)?)),
-		)
-		.expect("run attempt should exist");
-
-	assert!(blockers.contains(&String::from("thread_active")));
-	assert_eq!(thread_id, None);
-	assert_eq!(turn_id, None);
-}
-
-#[test]
-fn ghost_lane_cleanup_status_blockers_reject_existing_tracker_issue() {
-	let (_temp_dir, config, workflow) = running_lanes::temp_project_layout();
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let mut issue = running_lanes::sample_issue("In Progress", &[]);
-
-	issue.id = String::from("PUB-012");
-	issue.identifier = String::from("PUB-012");
-
-	let tracker = FakeTracker::new(vec![issue]);
-
-	state_store
-		.record_run_attempt("run-12", "PUB-012", 1, "running")
-		.expect("run attempt should record");
-	state_store
-		.upsert_lease("pubfi", "PUB-012", "run-12", "In Progress")
-		.expect("lease should record");
-
-	let blockers = orchestrator::ghost_lane_cleanup_status_blockers(
-		&tracker,
-		&config,
-		&workflow,
-		&state_store,
-		"PUB-012",
-		"run-12",
-	)
-	.expect("cleanup status blockers should load");
-
-	assert!(blockers.contains(&String::from("tracker_issue_present")));
-	assert!(blockers.contains(&String::from("issue_state:In Progress")));
-}
-
-#[test]
-fn ghost_lane_cleanup_status_blockers_reject_recent_protocol_evidence() {
-	let (_temp_dir, config, workflow) = running_lanes::temp_project_layout();
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let tracker = FakeTracker::new(Vec::new());
-	let worktree_path = config.worktree_root().join("PUB-012");
-
-	fs::create_dir_all(&worktree_path).expect("worktree path should exist");
-
-	state_store
-		.record_run_attempt("run-12", "PUB-012", 1, "running")
-		.expect("run attempt should record");
-	state_store
-		.upsert_lease("pubfi", "PUB-012", "run-12", "In Progress")
-		.expect("lease should record");
-	state_store
-		.upsert_worktree(
-			"pubfi",
-			"PUB-012",
-			"x/pubfi-pub-012",
-			&worktree_path.display().to_string(),
-		)
-		.expect("worktree should record");
-
-	state::write_run_protocol_activity_marker(
-		&worktree_path,
-		&ProtocolActivityMarker {
-			run_id: "run-12",
-			attempt_number: 1,
-			thread_id: None,
-			turn_id: None,
-			event_count: 1,
-			last_event_type: "thread/status/changed",
-			child_agent_activity: None,
-			protocol_activity: None,
-		},
-	)
-	.expect("recent protocol marker should write");
-	running_lanes::rewrite_run_activity_marker_host_boot_id(&worktree_path, "previous-boot");
-
-	let blockers = orchestrator::ghost_lane_cleanup_status_blockers(
-		&tracker,
-		&config,
-		&workflow,
-		&state_store,
-		"PUB-012",
-		"run-12",
-	)
-	.expect("cleanup status blockers should load");
-
-	assert!(blockers.contains(&String::from("protocol_recent")));
-	assert!(blockers.contains(&String::from("retained_worktree_present")));
-}
+mod blocker_lookup;
+mod live_evidence;
+mod status_snapshot;
