+++ apps/decodex/src/orchestrator/tests/operator/status/running_lanes/ghost_lane/cleanup_blockers/live_evidence.rs
@@ -1,0 +1,207 @@
+use crate::orchestrator::tests::operator::status::running_lanes::{
+	self, Connection, FakeTracker, ProtocolActivityMarker, StateStore, fs, orchestrator, process,
+	state,
+};
+
+#[test]
+fn ghost_lane_cleanup_status_blockers_reject_live_process_evidence() {
+	let (_temp_dir, config, workflow) = running_lanes::temp_project_layout();
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let tracker = FakeTracker::new(Vec::new());
+	let worktree_path = config.worktree_root().join("PUB-012");
+
+	fs::create_dir_all(&worktree_path).expect("worktree path should exist");
+
+	state_store
+		.record_run_attempt("run-12", "PUB-012", 1, "running")
+		.expect("run attempt should record");
+	state_store
+		.upsert_lease("pubfi", "PUB-012", "run-12", "In Progress")
+		.expect("lease should record");
+	state_store
+		.upsert_worktree(
+			"pubfi",
+			"PUB-012",
+			"x/pubfi-pub-012",
+			&worktree_path.display().to_string(),
+		)
+		.expect("worktree should record");
+
+	state::write_run_activity_marker_for_process(&worktree_path, "run-12", 1, process::id())
+		.expect("live process marker should write");
+
+	let blockers = orchestrator::ghost_lane_cleanup_status_blockers(
+		&tracker,
+		&config,
+		&workflow,
+		&state_store,
+		"PUB-012",
+		"run-12",
+	)
+	.expect("cleanup status blockers should load");
+
+	assert!(blockers.contains(&String::from("process_alive")));
+	assert!(blockers.contains(&String::from("retained_worktree_present")));
+}
+
+#[test]
+fn ghost_lane_cleanup_status_blockers_reject_active_thread_evidence() {
+	let (_temp_dir, config, workflow) = running_lanes::temp_project_layout();
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let tracker = FakeTracker::new(Vec::new());
+	let worktree_path = config.worktree_root().join("PUB-012");
+
+	fs::create_dir_all(&worktree_path).expect("worktree path should exist");
+
+	state_store
+		.record_run_attempt("run-12", "PUB-012", 1, "running")
+		.expect("run attempt should record");
+	state_store
+		.upsert_lease("pubfi", "PUB-012", "run-12", "In Progress")
+		.expect("lease should record");
+	state_store
+		.upsert_worktree(
+			"pubfi",
+			"PUB-012",
+			"x/pubfi-pub-012",
+			&worktree_path.display().to_string(),
+		)
+		.expect("worktree should record");
+
+	state::write_run_thread_status_marker(
+		&worktree_path,
+		"run-12",
+		1,
+		Some("thread-12"),
+		Some("turn-12"),
+		"active",
+		&[String::from("waitingOnApproval")],
+	)
+	.expect("active thread marker should write");
+
+	let blockers = orchestrator::ghost_lane_cleanup_status_blockers(
+		&tracker,
+		&config,
+		&workflow,
+		&state_store,
+		"PUB-012",
+		"run-12",
+	)
+	.expect("cleanup status blockers should load");
+
+	assert!(blockers.contains(&String::from("thread_active")));
+	assert!(blockers.contains(&String::from("retained_worktree_present")));
+}
+
+#[test]
+fn ghost_lane_cleanup_status_blockers_do_not_persist_marker_thread_identity() {
+	let (temp_dir, config, workflow) = running_lanes::temp_project_layout();
+	let state_path = temp_dir.path().join("runtime.sqlite3");
+	let state_store = StateStore::open(&state_path).expect("state store should open");
+	let tracker = FakeTracker::new(Vec::new());
+	let worktree_path = config.worktree_root().join("PUB-012");
+
+	fs::create_dir_all(&worktree_path).expect("worktree path should exist");
+
+	state_store
+		.record_run_attempt("run-12", "PUB-012", 1, "running")
+		.expect("run attempt should record");
+	state_store
+		.upsert_lease("pubfi", "PUB-012", "run-12", "In Progress")
+		.expect("lease should record");
+	state_store
+		.upsert_worktree(
+			"pubfi",
+			"PUB-012",
+			"x/pubfi-pub-012",
+			&worktree_path.display().to_string(),
+		)
+		.expect("worktree should record");
+
+	state::write_run_thread_status_marker(
+		&worktree_path,
+		"run-12",
+		1,
+		Some("thread-12"),
+		Some("turn-12"),
+		"active",
+		&[String::from("waitingOnApproval")],
+	)
+	.expect("active thread marker should write");
+
+	let blockers = orchestrator::ghost_lane_cleanup_status_blockers(
+		&tracker,
+		&config,
+		&workflow,
+		&state_store,
+		"PUB-012",
+		"run-12",
+	)
+	.expect("cleanup status blockers should load");
+	let connection = Connection::open(&state_path).expect("sqlite should open");
+	let (thread_id, turn_id): (Option<String>, Option<String>) = connection
+		.query_row(
+			"SELECT thread_id, turn_id FROM run_attempts WHERE run_id = 'run-12'",
+			[],
+			|row| Ok((row.get(0)?, row.get(1)?)),
+		)
+		.expect("run attempt should exist");
+
+	assert!(blockers.contains(&String::from("thread_active")));
+	assert_eq!(thread_id, None);
+	assert_eq!(turn_id, None);
+}
+
+#[test]
+fn ghost_lane_cleanup_status_blockers_reject_recent_protocol_evidence() {
+	let (_temp_dir, config, workflow) = running_lanes::temp_project_layout();
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let tracker = FakeTracker::new(Vec::new());
+	let worktree_path = config.worktree_root().join("PUB-012");
+
+	fs::create_dir_all(&worktree_path).expect("worktree path should exist");
+
+	state_store
+		.record_run_attempt("run-12", "PUB-012", 1, "running")
+		.expect("run attempt should record");
+	state_store
+		.upsert_lease("pubfi", "PUB-012", "run-12", "In Progress")
+		.expect("lease should record");
+	state_store
+		.upsert_worktree(
+			"pubfi",
+			"PUB-012",
+			"x/pubfi-pub-012",
+			&worktree_path.display().to_string(),
+		)
+		.expect("worktree should record");
+
+	state::write_run_protocol_activity_marker(
+		&worktree_path,
+		&ProtocolActivityMarker {
+			run_id: "run-12",
+			attempt_number: 1,
+			thread_id: None,
+			turn_id: None,
+			event_count: 1,
+			last_event_type: "thread/status/changed",
+			child_agent_activity: None,
+			protocol_activity: None,
+		},
+	)
+	.expect("recent protocol marker should write");
+	running_lanes::rewrite_run_activity_marker_host_boot_id(&worktree_path, "previous-boot");
+
+	let blockers = orchestrator::ghost_lane_cleanup_status_blockers(
+		&tracker,
+		&config,
+		&workflow,
+		&state_store,
+		"PUB-012",
+		"run-12",
+	)
+	.expect("cleanup status blockers should load");
+
+	assert!(blockers.contains(&String::from("protocol_recent")));
+	assert!(blockers.contains(&String::from("retained_worktree_present")));
+}
